Code refactoring pcs 19 (#29)
* Code refactoring for project entity

* Some feedback changes

* Changing the description size

* Local connection string

* refactoring changes and feedback modifications

* Little change in viewCreate Project

* Little change in Project Name link appearance

* Create or alter in store procedure for projectNameCheck

* Correct a bug for a project list

* Little changes in view

* Correct bugs in Project List View

* Bug changes for ProjectForm

* Delete extra code

* delete details

+++ Planilla_UCR/Domain/Projects/Entities/Project.cs
@@ -8,12 +8,12 @@ namespace Domain.Projects.Entities
         public String EmployerEmail { set; get; }
         public String ProjectName { set; get; }
         public String ProjectDescription { set; get; }
-        public int MaximumAmountForBenefits { set; get; }
+        public double MaximumAmountForBenefits { set; get; }
         public int MaximumBenefitAmount { set; get; }
         public String PaymentInterval { set; get; }
 
         public Project(String employerEmail, String projectName,
-                        String projectDescription, int maximumAmountForBenefits,
+                        String projectDescription, double maximumAmountForBenefits,
                         int maximumBenefitAmount, String paymentInterval)
         {
             EmployerEmail = employerEmail;
